A pipeline script has been created to automatically run all the process steps.
The pipeline is no longer speficied only for .wav stimuli. Durations, file list and associated conditions must be now written in a CSV file. This file is an input of the script of parameters file creation.

Some tools scripts and examples have been created. Somes are not finished yet.

The test directory will probably be removed soon.

Viewer needed somme modification to match the pipeline.